feat: implement the new design for the Views page
