+++ package.json
@@ -33,32 +33,31 @@
     "airtable": "^0.10.0",
     "eleventy-plugin-pwa": "^1.0.8",
     "eleventy-plugin-sass": "^1.1.1",
-    "infusion": "^3.0.0-dev.20210204T181946Z.3ed160b36.FLUID-6580",
-    "nodemailer": "^6.4.17",
+    "infusion": "^3.0.0-dev.20210223T161058Z.1743e0020.FLUID-6580",
+    "nodemailer": "^6.4.18",
     "sanitize.css": "^12.0.1",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-conventional": "^11.0.0",
+    "@commitlint/cli": "^12.0.0",
+    "@commitlint/config-conventional": "^12.0.0",
     "@tbranyen/jsdom": "^13.0.0",
     "autoprefixer": "^10.2.4",
     "axios": "^0.21.1",
     "eleventy-plugin-error-overlay": "^0.1.2",
-    "eslint": "^7.19.0",
+    "eslint": "^7.20.0",
     "html-minifier": "^4.0.0",
-    "husky": "^4.3.8",
+    "husky": "^5.1.1",
     "lint-staged": "^10.5.4",
     "markdown-it": "^12.0.4",
     "markdownlint-cli": "^0.26.0",
     "npm-run-all": "^4.1.5",
     "release": "^6.3.0",
     "rimraf": "^3.0.2",
-    "slugify": "^1.4.5",
-    "stylelint": "^13.9.0",
+    "stylelint": "^13.11.0",
     "stylelint-config-standard-scss": "^1.1.0",
     "stylelint-order": "^4.1.0",
-    "stylelint-scss": "^3.18.0"
+    "stylelint-scss": "^3.19.0"
   },
   "husky": {
     "hooks": {
