--- src/_includes/components/workshop.njk
@@ -1,15 +1,0 @@
-<a href="/initiatives/{{ workshop.id }}">
-	<div class="one-workshop">
-		<div class="workshop-narrative">
-			<h2 class="h3">{{ workshop.title | safe }}</h2>
-			<div class="description">{{ workshop.shortDescription | safe }}</div>
-		</div>
-		<div class="workshop-image">
-			{% if workshop.previewImageUrl %}
-				<img src="{{ workshop.previewImageUrl | safe }}" {% if workshop.previewImageAltText %}alt="{{ workshop.previewImageAltText}}"{% else %}role="presentation"{% endif %}>
-			{% else %}
-				<img src="/images/workshop.png" role="presentation">
-			{% endif %}
-		</div>
-	</div>
-</a>
