+++ src/scss/abstracts/_mixins.scss
@@ -157,8 +157,8 @@
 		}
 	}
 
-	// Initiatives page: Workshop section
-	.one-workshop {
+	// Tile item
+	.tile-item {
 		box-shadow: rem(0) rem(0) rem(7) currentColor;
 
 		&:hover {
