+++ src/scss/components/_components.scss
@@ -7,7 +7,9 @@
 @import "filter";
 @import "initiatives";
 @import "news-and-views";
+@import "news-grid";
 @import "pagination";
 @import "post-article";
 @import "search-container";
 @import "team";
+@import "tile-item";
