+++ src/scss/components/_news-grid.scss
@@ -1,0 +1,93 @@
+.news-grid {
+	display: grid;
+	grid-gap: rem(30) rem(38);
+	grid-template-columns: repeat(1, 1fr);
+
+	.api-content {
+		background: #fff;
+		border: 1px solid #000;
+		box-sizing: border-box;
+
+		.news-item {
+			.preview-media-wrapper .content {
+				display: none;
+			}
+
+			.author {
+				color: $grey-light;
+				font-style: italic;
+				font-weight: $font-weight-semibold;
+			}
+
+			.date {
+				padding-bottom: rem(8);
+			}
+
+			// Apply ellipsis on the third line to the preview content.
+			// Reference: http://hackingui.com/front-end/a-pure-css-solution-for-multiline-text-truncation/
+			.preview-content {
+				line-height: rem(19.2);
+				margin-bottom: rem(16);
+				margin-right: rem(-16);
+				max-height: rem(57.6);
+				overflow: hidden;
+				padding-right: rem(16);
+				position: relative;
+			}
+
+			// create the ellipsis
+			.preview-content::before {
+				bottom: 0;
+				content: '...';
+				position: absolute;
+				right: 0;
+			}
+
+			// hide the ellipsis if we have text, which is less than or equal to max lines
+			.preview-content::after {
+				background: $white;
+				content: '';
+				height: rem(16);
+				margin-top: rem(3.2);
+				position: absolute;
+				right: 0;
+				width: rem(16);
+			}
+		}
+	}
+}
+
+@include breakpoint-up(lg) {
+	.news-grid {
+		grid-template-columns: repeat(2, 1fr);
+
+		.api-content .news-item {
+			padding: 0;
+
+			.preview-media-wrapper {
+				@include aspect-ratio(16, 9);
+
+				.content {
+					display: block;
+				}
+
+				// Style the placeholder images
+				.placeholder-img {
+					height: 100%;
+					width: 100%;
+				}
+
+				// Style featured images
+				img {
+					height: 100%;
+					object-fit: cover;
+					width: 100%;
+				}
+			}
+		}
+
+		.content-preview {
+			padding: rem(1) rem(60);
+		}
+	}
+}
