+++ src/scss/components/_tile-item.scss
@@ -1,0 +1,73 @@
+.tile-item {
+	border-radius: rem(18);
+	box-shadow: rem(0) rem(0) rem(7) rgba(0, 0, 0, 0.25);
+	display: grid;
+	margin: rem(32) 0 rem(32) 0;
+
+	&:hover {
+		box-shadow: 0 0 rem(35) rgba(0, 0, 0, 0.55);
+	}
+
+	// Workshop title and descriptions
+	.tile-narrative {
+		border-radius: rem(18) rem(18) rem(18) rem(18);
+		min-height: rem(208);
+		padding: 0 rem(16) rem(16) rem(16);
+
+		// Workshop titles
+		h2 {
+			border-bottom: rem(2) solid $grey-shade;
+			padding-bottom: rem(16);
+		}
+
+		.author {
+			color: $black;
+			font-style: italic;
+			font-weight: $font-weight-semibold;
+		}
+
+		.date {
+			color: $grey-light;
+			font-style: italic;
+			padding-bottom: rem(16);
+		}
+
+		.description {
+			color: $black;
+			font-weight: $font-weight-normal;
+		}
+	}
+
+	// Workshop images
+	.tile-image {
+		display: none;
+	}
+}
+
+@include breakpoint-up(lg) {
+	.tile-item {
+		grid-template-columns: 70% 30%;
+		grid-template-rows: 100%;
+
+		.tile-narrative {
+			border-radius: rem(18) 0 0 rem(18);
+			min-height: rem(336);
+			padding: 0 rem(16) rem(16) rem(32);
+		}
+
+		// In order to apply border-radius, all inner elements need to have the same border-radius applied.
+		// This includes the wrapper div.tile-image and inner img element.
+		.tile-image {
+			border-radius: 0 0 rem(18) rem(18);
+			display: block;
+			height: 100%;
+
+			img,
+			svg {
+				border-radius: 0 rem(18) rem(18) 0;
+				height: 100%;
+				width: 100%;
+			}
+		}
+	}
+}
